+++ project/__init__.py
@@ -1,1 +1,8 @@
 
+from project.baseproject import BaseProject
+from project.remoteproject import RemoteProject
+
+__all__ = (
+    BaseProject,
+    RemoteProject,
+)
